вынес сортировку отдельно
